Create package protocols/echo. Move echo handler.
