+++ src/protocols/echo.rs
@@ -1,0 +1,19 @@
+use std::io;
+use futures::{AsyncReadExt, AsyncWriteExt};
+use libp2p::Stream;
+
+pub async fn handler(mut stream: Stream) -> io::Result<usize> {
+    let mut total = 0;
+
+    let mut buf = [0u8; 100];
+
+    loop {
+        let read = stream.read(&mut buf).await?;
+        if read == 0 {
+            return Ok(total);
+        }
+
+        total += read;
+        stream.write_all(&buf[..read]).await?;
+    }
+}
